+++ packages/lightspark-cli/.eslintrc.cjs
@@ -1,3 +1,4 @@
 module.exports = {
   extends: ["@lightsparkdev/eslint-config/base"],
+  ignorePatterns: ["lightspark_crypto.js"],
 };