+++ packages/lightspark-cli/src/index.ts
@@ -12,12 +12,18 @@ import type { OptionValues } from "commander";
 import { Command, InvalidArgumentError } from "commander";
 import * as fs from "fs/promises";
 import qrcode from "qrcode-terminal";
+import {
+  LightsparkSigner,
+  Mnemonic,
+  Seed,
+} from "../lightspark_crypto/lightspark_crypto.js";
 import type { EnvCredentials } from "./authHelpers.js";
 import { getCredentialsFromEnvOrThrow } from "./authHelpers.js";
 import {
-  getBitcoinNetworkOrDefault,
+  getBitcoinNetworkOrThrow,
   getNodeId,
   getPackageVersion,
+  isBitcoinNetwork,
 } from "./helpers.js";
 
 const main = async (
@@ -25,7 +31,7 @@ const main = async (
   action: (
     client: LightsparkClient,
     options: OptionValues,
-    credentials?: EnvCredentials
+    credentials: EnvCredentials
   ) => Promise<unknown>
 ) => {
   const credentials = getCredentialsFromEnvOrThrow();
@@ -76,24 +82,54 @@ const initEnv = async (options: OptionValues) => {
 
   let content = `export LIGHTSPARK_API_TOKEN_CLIENT_ID="${clientId}"\n`;
   content += `export LIGHTSPARK_API_TOKEN_CLIENT_SECRET="${clientSecret}"\n`;
+  let baseUrl: string | undefined;
   if (options.env === "dev") {
-    content += `export LIGHTSPARK_EXAMPLE_BASE_URL="api.dev.dev.sparkinfra.net"\n`;
+    baseUrl = "api.dev.dev.sparkinfra.net";
+    content += `export LIGHTSPARK_EXAMPLE_BASE_URL="${baseUrl}"\n`;
   }
+
+  const client = new LightsparkClient(
+    new AccountTokenAuthProvider(clientId, clientSecret),
+    baseUrl
+  );
+
+  let tokenBitcoinNetwork;
+  for (const bitcoinNetwork of Object.values(BitcoinNetwork)) {
+    try {
+      await client.getAccountDashboard(undefined, bitcoinNetwork);
+
+      console.log(`API token is for bitcoin network ${bitcoinNetwork}`);
+      tokenBitcoinNetwork = bitcoinNetwork;
+    } catch (e) {
+      // Do nothing
+    }
+  }
+
+  if (!tokenBitcoinNetwork) {
+    console.log(
+      "Could not find any nodes on the account with the provided API token client ID and secret for any bitcoin network."
+    );
+    tokenBitcoinNetwork = await input({
+      message:
+        "Please provide the bitcoin network associated with this API token (test -> REGTEST, prod -> MAINNET)?",
+      validate: (value) => isBitcoinNetwork(value as BitcoinNetwork),
+    });
+  }
+
+  content += `export BITCOIN_NETWORK="${tokenBitcoinNetwork}"`;
+
   await fs.writeFile(filePath, content);
 
   console.log("Wrote environment variables to " + filePath);
-  console.log("Run `source " + filePath + "` to load them into your shell");
-  console.log(
-    "To add them to your shell permanently, add the above line to your shell's startup script"
-  );
   console.log("You can now run `lightspark` to interact with your wallet");
 };
 
 const createInvoice = async (
   client: LightsparkClient,
-  options: OptionValues
+  options: OptionValues,
+  credentials: EnvCredentials
 ) => {
-  const bitcoinNetwork = getBitcoinNetworkOrDefault(options);
+  const bitcoinNetwork = getBitcoinNetworkOrThrow(credentials.bitcoinNetwork);
   const nodeId = await getNodeId(client, bitcoinNetwork);
 
   let { amount, memo, amp } = options;
@@ -225,9 +261,10 @@ const l1FeeEstimate = async (
 
 const accountDashboard = async (
   client: LightsparkClient,
-  options: OptionValues
+  options: OptionValues,
+  credentials: EnvCredentials
 ) => {
-  const bitcoinNetwork = getBitcoinNetworkOrDefault(options);
+  const bitcoinNetwork = getBitcoinNetworkOrThrow(credentials.bitcoinNetwork);
 
   console.log(`Fetching account dashboard for ${bitcoinNetwork}...\n`);
   const dashboard = await client.getAccountDashboard(undefined, bitcoinNetwork);
@@ -236,9 +273,10 @@ const accountDashboard = async (
 
 const singleNodeDashboard = async (
   client: LightsparkClient,
-  options: OptionValues
+  options: OptionValues,
+  credentials: EnvCredentials
 ) => {
-  const bitcoinNetwork = getBitcoinNetworkOrDefault(options);
+  const bitcoinNetwork = getBitcoinNetworkOrThrow(credentials.bitcoinNetwork);
   const nodeId = await getNodeId(client, bitcoinNetwork);
 
   console.log(`Fetching single node dashboard in ${bitcoinNetwork}...\n`);
@@ -273,9 +311,10 @@ const decodeInvoice = async (
 
 const createNodeWalletAddress = async (
   client: LightsparkClient,
-  options: OptionValues
+  options: OptionValues,
+  credentials: EnvCredentials
 ) => {
-  const bitcoinNetwork = getBitcoinNetworkOrDefault(options);
+  const bitcoinNetwork = getBitcoinNetworkOrThrow(credentials.bitcoinNetwork);
   const nodeId = await getNodeId(client, bitcoinNetwork);
 
   console.log(`Creating bitcoin funding address for node ${nodeId}...\n`);
@@ -286,9 +325,9 @@ const createNodeWalletAddress = async (
 const payInvoice = async (
   client: LightsparkClient,
   options: OptionValues,
-  credentials?: EnvCredentials
+  credentials: EnvCredentials
 ) => {
-  const bitcoinNetwork = getBitcoinNetworkOrDefault(options);
+  const bitcoinNetwork = getBitcoinNetworkOrThrow(credentials.bitcoinNetwork);
   const nodeId = await getNodeId(client, bitcoinNetwork);
 
   let encodedInvoice = options.invoice;
@@ -300,8 +339,8 @@ const payInvoice = async (
   }
 
   let nodePassword = options.password;
-  if (bitcoinNetwork === BitcoinNetwork.MAINNET) {
-    nodePassword = nodePassword || credentials?.testNodePassword;
+  if (bitcoinNetwork === BitcoinNetwork.REGTEST) {
+    nodePassword = nodePassword || credentials.testNodePassword;
   } else if (!nodePassword) {
     nodePassword = await password({
       message: `What is the password for the node (${nodeId})?`,
@@ -358,9 +397,10 @@ const createTestModePayment = async (
 
 const getNodeChannels = async (
   client: LightsparkClient,
-  options: OptionValues
+  options: OptionValues,
+  credentials: EnvCredentials
 ) => {
-  const bitcoinNetwork = getBitcoinNetworkOrDefault(options);
+  const bitcoinNetwork = getBitcoinNetworkOrThrow(credentials.bitcoinNetwork);
 
   const account = await client.getCurrentAccount();
   console.log("Got account:", JSON.stringify(account, null, 2));
@@ -370,6 +410,46 @@ const getNodeChannels = async (
 
   const channels = (await nodes.entities[0]?.getChannels(client, 20)) || {};
   console.log("Got channels:", JSON.stringify(channels, null, 2));
+};
+
+const generateNodeKeys = async (
+  client: LightsparkClient,
+  options: OptionValues
+) => {
+  let mnemonic: Mnemonic;
+  let seedPhrase = options.seedPhrase;
+  if (!seedPhrase) {
+    const isProvidingSeedPhrase = await input({
+      message: "Provide your own seed phrase (Y/n)?",
+      validate: (value) =>
+        value.toUpperCase() === "Y" || value.toUpperCase() === "N",
+    });
+
+    if (isProvidingSeedPhrase.toUpperCase() === "Y") {
+      seedPhrase = await input({
+        message: "What is your seed phrase (ex: horse battery staple ...)?",
+        validate: (value) => value.split(" ").length === 24,
+      });
+      mnemonic = Mnemonic.from_phrase(seedPhrase);
+    } else {
+      mnemonic = Mnemonic.new();
+    }
+  } else {
+    mnemonic = Mnemonic.from_phrase(seedPhrase);
+  }
+
+  console.log("Generating node keys...\n");
+
+  const seed = Seed.from_mnemonic(mnemonic);
+  const lightsparkSigner = LightsparkSigner.new(seed);
+  const extendedPublicKey = lightsparkSigner.get_master_public_key();
+
+  const seedAsHex = seed.as_bytes().reduce((acc: string, byte: number) => {
+    return (acc += ("0" + byte.toString(16)).slice(-2));
+  }, "");
+  console.log(`Seed phrase:\n${mnemonic.as_string()}\n`);
+  console.log(`Hex-encoded seed:\n${seedAsHex}\n`);
+  console.log(`Extended public key:\n${extendedPublicKey}`);
 };
 
 const safeParseInt = (value: string /* dummyPrevious: any */) => {
@@ -384,10 +464,6 @@ const safeParseInt = (value: string /* dummyPrevious: any */) => {
 (() => {
   const createInvoiceCmd = new Command("create-invoice")
     .description("Create an invoice for your account")
-    .option(
-      "-b --bitcoin-network <value>",
-      "The bitcoin network for which to get a node's channels. Defaults to MAINNET."
-    )
     .option(
       "-m, --memo  <value>",
       "Add a memo describing the invoice.",
@@ -408,10 +484,6 @@ const safeParseInt = (value: string /* dummyPrevious: any */) => {
 
   const createTestModeInvoiceCmd = new Command("create-test-mode-invoice")
     .description("Create an invoice that you can pay in test mode.")
-    .option(
-      "-b --bitcoin-network <value>",
-      "The bitcoin network for which to get a node's channels. Defaults to MAINNET."
-    )
     .option(
       "-m, --memo  <value>",
       "Add a memo describing the invoice.",
@@ -484,10 +556,6 @@ const safeParseInt = (value: string /* dummyPrevious: any */) => {
 
   const accountDashboardCmd = new Command("account-dashboard")
     .description("Get account dashboard")
-    .option(
-      "-b --bitcoin-network <value>",
-      "The bitcoin network to include in the dashboard. Defaults to MAINNET."
-    )
     .action((options) => {
       main(options, accountDashboard).catch((err) =>
         console.error("Oh no, something went wrong.\n", err)
@@ -496,10 +564,6 @@ const safeParseInt = (value: string /* dummyPrevious: any */) => {
 
   const singleNodeDashboardCmd = new Command("single-node-dashboard")
     .description("Get a single node's dashboard")
-    .option(
-      "-b --bitcoin-network <value>",
-      "The bitcoin network for which to get a node's dashboard. Defaults to MAINNET."
-    )
     .action((options) => {
       main(options, singleNodeDashboard).catch((err) =>
         console.error("Oh no, something went wrong.\n", err)
@@ -516,10 +580,6 @@ const safeParseInt = (value: string /* dummyPrevious: any */) => {
 
   const createNodeWalletAddressCmd = new Command("funding-address")
     .description("Create a bitcoin funding address for your account.")
-    .option(
-      "-b --bitcoin-network <value>",
-      "The bitcoin network for which to get a node's channels. Defaults to MAINNET."
-    )
     .action((options) => {
       main(options, createNodeWalletAddress).catch((err) =>
         console.error("Oh no, something went wrong.\n", err)
@@ -528,10 +588,6 @@ const safeParseInt = (value: string /* dummyPrevious: any */) => {
 
   const payInvoiceCmd = new Command("pay-invoice")
     .description("Pay an invoice from your account.")
-    .option(
-      "-b --bitcoin-network <value>",
-      "The bitcoin network for which to get a node's channels. Defaults to MAINNET."
-    )
     .option("-i, --invoice  <value>", "The encoded payment request.")
     .option(
       "-a, --amount <number>",
@@ -549,10 +605,6 @@ const safeParseInt = (value: string /* dummyPrevious: any */) => {
     .description(
       "In test mode, simulates a payment from another node to an invoice."
     )
-    .option(
-      "-b --bitcoin-network <value>",
-      "The bitcoin network for which to get a node's channels. Defaults to MAINNET."
-    )
     .option("-i, --invoice  <value>", "The encoded payment request.")
     .option(
       "-a, --amount <number>",
@@ -568,12 +620,20 @@ const safeParseInt = (value: string /* dummyPrevious: any */) => {
 
   const getNodeChannelsCmd = new Command("node-channels")
     .description("Gets node channels.")
-    .option(
-      "-b --bitcoin-network <value>",
-      "The bitcoin network for which to get a node's channels. Defaults to MAINNET."
-    )
     .action((options) => {
       main(options, getNodeChannels).catch((err) =>
+        console.error("Oh no, something went wrong.\n", err)
+      );
+    });
+
+  const generateNodeKeysCmd = new Command("generate-node-keys")
+    .description("Generates keys needed for your Lightspark node.")
+    .option(
+      "-s --seed-phrase <value>",
+      "The seed phrase used to generate the keys need for your Lightspark node."
+    )
+    .action((options) => {
+      main(options, generateNodeKeys).catch((err) =>
         console.error("Oh no, something went wrong.\n", err)
       );
     });
@@ -608,6 +668,7 @@ const safeParseInt = (value: string /* dummyPrevious: any */) => {
     .addCommand(payInvoiceCmd)
     .addCommand(createTestModePaymentCmd)
     .addCommand(getNodeChannelsCmd)
+    .addCommand(generateNodeKeysCmd)
     .addCommand(InitEnvCmd)
     .addHelpCommand()
     .parse(process.argv);