+++ packages/ui/components/index.ts
@@ -4,11 +4,9 @@ export {
   StyledButton,
   type ButtonProps,
 } from "./Button";
-export {
-  CurrencyAmount,
-  CurrencyIcon,
-  type CurrencyAmountType,
-} from "./CurrencyAmount";
+export { CopyToClipboardButton } from "./CopyToClipboardButton";
+export { CurrencyAmount, CurrencyIcon } from "./CurrencyAmount";
 export { Icon, IconContainer } from "./Icon";
 export { Loading, LoadingWrapper } from "./Loading";
+export { SecretContainer } from "./SecretContainer";
 export { UnstyledButton } from "./UnstyledButton";