+++ packages/wallet-sdk/src/objects/IncomingPayment.ts
@@ -1,6 +1,6 @@
 // Copyright ©, 2023-present, Lightspark Group, Inc. - All Rights Reserved
 
-import type { Query } from "@lightsparkdev/core";
+import { type Query } from "@lightsparkdev/core";
 import type CurrencyAmount from "./CurrencyAmount.js";
 import { CurrencyAmountFromJson } from "./CurrencyAmount.js";
 import type Entity from "./Entity.js";
@@ -8,7 +8,7 @@ import type LightningTransaction from "./LightningTransaction.js";
 import type Transaction from "./Transaction.js";
 import TransactionStatus from "./TransactionStatus.js";
 
-/** A transaction that was sent to a Lightspark node on the Lightning Network. **/
+/** This object represents any payment sent to a Lightspark node on the Lightning Network. You can retrieve this object to receive payment related information about a specific payment received by a Lightspark node. **/
 type IncomingPayment = LightningTransaction &
   Transaction &
   Entity & {