+++ packages/wallet-sdk/src/objects/Node.ts
@@ -1,7 +1,6 @@
 // Copyright ©, 2023-present, Lightspark Group, Inc. - All Rights Reserved
 
-import type { Query } from "@lightsparkdev/core";
-import { LightsparkException } from "@lightsparkdev/core";
+import { LightsparkException, type Query } from "@lightsparkdev/core";
 import autoBind from "auto-bind";
 import type LightsparkClient from "../client.js";
 import BitcoinNetwork from "./BitcoinNetwork.js";
@@ -11,7 +10,7 @@ import type NodeAddressType from "./NodeAddressType.js";
 import type NodeToAddressesConnection from "./NodeToAddressesConnection.js";
 import { NodeToAddressesConnectionFromJson } from "./NodeToAddressesConnection.js";
 
-/** This interface represents a lightning node that can be connected to the Lightning Network to send and receive transactions. **/
+/** This object is an interface representing a Lightning Node on the Lightning Network, and could either be a Lightspark node or a node managed by a third party. **/
 class Node implements Entity {
   constructor(
     public readonly id: string,