+++ packages/wallet-sdk/src/objects/OutgoingPayment.ts
@@ -1,6 +1,6 @@
 // Copyright ©, 2023-present, Lightspark Group, Inc. - All Rights Reserved
 
-import type { Query } from "@lightsparkdev/core";
+import { type Query } from "@lightsparkdev/core";
 import type CurrencyAmount from "./CurrencyAmount.js";
 import { CurrencyAmountFromJson } from "./CurrencyAmount.js";
 import type Entity from "./Entity.js";
@@ -13,7 +13,7 @@ import { RichTextFromJson } from "./RichText.js";
 import type Transaction from "./Transaction.js";
 import TransactionStatus from "./TransactionStatus.js";
 
-/** A transaction that was sent from a Lightspark node on the Lightning Network. **/
+/** This object represents a Lightning Network payment sent from a Lightspark Node. You can retrieve this object to receive payment related information about any payment sent from your Lightspark Node on the Lightning Network. **/
 type OutgoingPayment = LightningTransaction &
   Transaction &
   Entity & {