+++ packages/wallet-sdk/src/objects/Transaction.ts
@@ -1,7 +1,6 @@
 // Copyright ©, 2023-present, Lightspark Group, Inc. - All Rights Reserved
 
-import type { Query } from "@lightsparkdev/core";
-import { LightsparkException } from "@lightsparkdev/core";
+import { LightsparkException, type Query } from "@lightsparkdev/core";
 import type ChannelClosingTransaction from "./ChannelClosingTransaction.js";
 import type ChannelOpeningTransaction from "./ChannelOpeningTransaction.js";
 import type CurrencyAmount from "./CurrencyAmount.js";
@@ -16,6 +15,7 @@ import { RichTextFromJson } from "./RichText.js";
 import TransactionStatus from "./TransactionStatus.js";
 import type Withdrawal from "./Withdrawal.js";
 
+/** This object represents a payment transaction. The transaction can occur either on a Bitcoin Network, or over the Lightning Network. You can retrieve this object to receive specific information about a particular transaction tied to your Lightspark Node. **/
 type Transaction = Entity & {
   /**
    * The unique identifier of this entity across all Lightspark systems. Should be treated as an opaque