+++ packages/lightspark-cli/CHANGELOG.md
@@ -1,15 +1,46 @@
 # @lightsparkdev/lightspark-cli
 
-## 0.0.1
+## 0.0.8
 
 ### Patch Changes
 
-- df28af1: Move tsconfig to devDependencies
+- 451b106: Output hex-encoded seed in generate-node-keys command
+
+## 0.0.7
+
+### Patch Changes
+
+- 5bf68c1: Fix parsing of webhook subevent type
+- 5bf68c1: Automatically find bitcoin network for provided api token
+- Updated dependencies [6cd80d4]
+  - @lightsparkdev/lightspark-sdk@0.4.9
+
+## 0.0.6
+
+### Patch Changes
+
+- 88a8994: Fix version retrieval for global installs
+
+## 0.0.5
+
+### Patch Changes
+
+- 0b71a9c: Add generate-node-key command
+
+## 0.0.4
+
+### Patch Changes
+
+- 565f0ab: Fix tsconfigs for CLIs to ensure builds output to dist
+
+## 0.0.3
+
+### Patch Changes
+
+- cf11549: Add ability to set the invoice expiration in createInvoice.
 
 ## 0.0.2
 
 ### Patch Changes
 
-- Updated dependencies [141d73a]
-  - @lightsparkdev/lightspark-sdk@0.4.8
-  - @lightsparkdev/core@0.3.10
+- 7423c87: Empty change to trigger version