+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @lightsparkdev/ui
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [6cd80d4]
+  - @lightsparkdev/wallet-sdk@0.6.16
+
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies [565f0ab]
+  - @lightsparkdev/wallet-sdk@0.6.15
+
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [cf11549]
+  - @lightsparkdev/wallet-sdk@0.6.14
+
 ## 0.0.3
 
 ### Patch Changes